pyproject.toml
[tool.poetry]
name = "random-gh"
version = "0.1.0"
description = ""
authors = ["chisun <csjoung@gmail.com>"]
readme = "README.md"
packages = [{include = "random_gh"}]

[tool.poetry.dependencies]
python = "^3.10"
PyGithub = "^1.58.1"
python-dotenv = "^1.0.0"
termcolor = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
